feat: 상품 등록, 수정, 삭제는 회원가입된 멤버만 가능하도록 변경

+++ src/main/java/gift/controller/ProductController.java
@@ -2,9 +2,18 @@ package gift.controller;
 
 import gift.dto.CreateProductRequestDto;
 import gift.dto.ProductResponseDto;
+import gift.entity.Member;
+import gift.exception.CustomException;
+import gift.exception.ErrorCode;
 import gift.service.ProductService;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
 import jakarta.validation.Valid;
 import java.util.List;
+import java.util.Optional;
+import javax.crypto.SecretKey;
+import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,6 +24,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,13 +34,20 @@ public class ProductController {
 
     private final ProductService productService;
 
+    private final String key = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E";
+
     public ProductController(ProductService productService) {
         this.productService = productService;
     }
 
     @PostMapping
     public ResponseEntity<ProductResponseDto> createProduct(
-            @Valid @RequestBody CreateProductRequestDto requestDto) {
+            @Valid @RequestBody CreateProductRequestDto requestDto,
+            @RequestHeader(value = "Authorization") String token) {
+        Optional<Member> optionalMember = isValidateToken(token);
+        if (optionalMember.isEmpty()) {
+            throw new CustomException(ErrorCode.NotLogin);
+        }
         return new ResponseEntity<>(productService.createProduct(requestDto), HttpStatus.CREATED);
     }
 
@@ -45,15 +62,44 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ProductResponseDto> updateProductById(@PathVariable Long id,
-            @Valid @RequestBody CreateProductRequestDto requestDto) {
+    public ResponseEntity<ProductResponseDto> updateProductById(
+            @PathVariable Long id,
+            @Valid @RequestBody CreateProductRequestDto requestDto,
+            @RequestHeader(value = "Authorization") String token) {
+        Optional<Member> optionalMember = isValidateToken(token);
+        if (optionalMember.isEmpty()) {
+            throw new CustomException(ErrorCode.NotLogin);
+        }
+
         return new ResponseEntity<>(productService.updateProductById(id, requestDto),
                 HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProductById(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProductById(
+            @PathVariable Long id,
+            @RequestHeader(value = "Authorization") String token) {
+        Optional<Member> optionalMember = isValidateToken(token);
+        if (optionalMember.isEmpty()) {
+            throw new CustomException(ErrorCode.NotLogin);
+        }
         productService.deleteProductById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    private Optional<Member> isValidateToken(String token) {
+        try {
+            Claims claims = Jwts.parser()
+                    .verifyWith(Keys.hmacShaKeyFor(key.getBytes()))
+                    .build()
+                    .parseSignedClaims(token)
+                    .getPayload();
+            Member find = new Member(null, claims.get("email", String.class), null,
+                    claims.get("role", String.class));
+            return Optional.of(find);
+
+        } catch (Exception e) {
+            return Optional.empty();
+        }
     }
 }
